refactor: move up AccountAction and AccountCreator
