+++ store/src/test/java/com/zimbra/cs/service/FileUploadServletTest.java
@@ -10,7 +10,7 @@ import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.mock;
 
 import com.google.common.collect.Maps;
-import com.zextras.mailbox.util.MailboxTestUtil.AccountCreator;
+import com.zextras.mailbox.util.AccountCreator;
 import com.zextras.mailbox.util.PortUtil;
 import com.zimbra.common.account.Key;
 import com.zimbra.common.account.ZAttrProvisioning;
