+++ store/src/test/java/com/zimbra/cs/service/account/GetInfoTest.java
@@ -1,6 +1,6 @@
 package com.zimbra.cs.service.account;
 
-import static com.zextras.mailbox.util.MailboxTestUtil.AccountCreator;
+import com.zextras.mailbox.util.AccountCreator;
 import static org.junit.jupiter.api.Assertions.*;
 
 import com.zextras.mailbox.soap.SoapTestSuite;
