+++ store/src/test/java/com/zimbra/cs/service/admin/ModifyDomainTest.java
@@ -6,7 +6,7 @@ package com.zimbra.cs.service.admin;
 
 
 import com.zextras.mailbox.soap.SoapTestSuite;
-import com.zextras.mailbox.util.MailboxTestUtil.AccountCreator;
+import com.zextras.mailbox.util.AccountCreator;
 import com.zimbra.common.account.ZAttrProvisioning;
 import com.zimbra.cs.account.Account;
 import com.zimbra.cs.account.Domain;
