+++ store/src/test/java/com/zimbra/cs/service/mail/CreateCalendarItemExceptionTest.java
@@ -3,7 +3,7 @@ package com.zimbra.cs.service.mail;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import com.zextras.mailbox.soap.SoapTestSuite;
-import com.zextras.mailbox.util.MailboxTestUtil.AccountCreator;
+import com.zextras.mailbox.util.AccountCreator;
 import com.zimbra.common.service.ServiceException;
 import com.zimbra.common.soap.Element;
 import com.zimbra.common.soap.Element.JSONElement;
