+++ store/src/test/java/com/zimbra/cs/service/mail/DeleteCalendarTest.java
@@ -1,6 +1,6 @@
 package com.zimbra.cs.service.mail;
 
-import static com.zextras.mailbox.util.MailboxTestUtil.AccountCreator;
+import com.zextras.mailbox.util.AccountCreator;
 import static com.zimbra.common.soap.Element.parseXML;
 import static org.junit.jupiter.api.Assertions.*;
